+++ addons/main/config.cpp
@@ -4,22 +4,36 @@ class CfgPatches {
     class ADDON {
         name = COMPONENT_NAME;
         units[] = {};
-        weapons[] = {};
+        weapons[] = {"LMG_Zafir_B_F"};
         requiredVersion = REQUIRED_VERSION;
         requiredAddons[] = {
-            "cba_main",
             "ace_realisticnames",
-            "A3_Characters_F",
-            "A3_Characters_F_beta",
-            "A3_Characters_F_epa",
-            "A3_Characters_F_epb",
-            "A3_Characters_F_epc",
-            "A3_Characters_F_exp"
+            "A3_Data_F_AoW_Loadorder"
         };
+        author = "johnb43";
         authors[] = {
             "johnb43",
             "Blutze"
         };
+        url = "https://github.com/johnb432/Weapons-Balance-Vanilla";
+        VERSION_CONFIG;
+    };
+};
+
+class CfgMods {
+    class PREFIX {
+        name = "Weapons Balance - Vanilla";
+        author = "johnb43";
+        tooltipOwned = "Weapons Balance - Vanilla";
+        hideName = 0;
+        hidePicture = 0;
+        actionName = "Github";
+        action = "https://github.com/johnb432/Weapons-Balance-Vanilla";
+        description = "A collection of balancing mods made by johnb43.";
+        overview = "A collection of balancing mods made by johnb43.";
+        picture = "\z\wb_vanilla\addons\main\ui\logo_weapons_balance.paa"; // http://getdrawings.com/get-drawing#gun-drawing-in-pencil-27.jpg, http://getdrawings.com/get-drawing#gun-drawing-in-pencil-17.png
+        logo = "\z\wb_vanilla\addons\main\ui\logo_weapons_balance.paa";
+        overviewPicture = "\z\wb_vanilla\addons\main\ui\logo_weapons_balance.paa";
     };
 };