+++ mod.cpp
@@ -1,5 +1,12 @@
 name = "Weapons Balance - Vanilla";
-description = "A collection of balancing mods made by johnb43";
-tooltip = "Weapons Balance - Vanilla";
-tooltipOwned = "Weapons Balance - Vanilla";
 author = "johnb43";
+tooltipOwned = "Weapons Balance - Vanilla";
+hideName = 0;
+hidePicture = 0;
+actionName = "Github";
+action = "https://github.com/johnb432/Weapons-Balance-Vanilla";
+description = "A collection of balancing mods made by johnb43.";
+overview = "A collection of balancing mods made by johnb43.";
+picture = "\z\wb_vanilla\addons\main\ui\logo_weapons_balance.paa"; // http://getdrawings.com/get-drawing#gun-drawing-in-pencil-27.jpg, http://getdrawings.com/get-drawing#gun-drawing-in-pencil-17.png
+logo = "\z\wb_vanilla\addons\main\ui\logo_weapons_balance.paa";
+overviewPicture = "\z\wb_vanilla\addons\main\ui\logo_weapons_balance.paa";